requirements.txt
pandas
openpyxl
python-docx
pdfplumber
pywin32
